third_party/blink/web_tests/dom/mutation-event-tests/fast/events/mutation-during-insert-before-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeRemoved' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Ensures that insertBefore() throws an exception if mutation even handler does something wrong

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS newparent.insertBefore(child, beforeChild); threw exception HierarchyRequestError: Failed to execute 'insertBefore' on 'Node': The new child element contains the parent..
PASS successfullyParsed is true

TEST COMPLETE
